requirements.txt
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
feedparser==6.0.10
requests==2.31.0
openai==1.3.0
sqlalchemy==2.0.23
apscheduler==3.10.4
python-dotenv==1.0.0
lxml==4.9.3
gunicorn==21.2.0
psycopg2-binary==2.9.9

